rename: Rating 컴포넌트 UserItem 하위로 이동 (#12)
