+++ src/components/UserItem/Rating/Rating.module.scss
