+++ src/components/UserItem/Rating/Rating.tsx
