+++ src/components/UserItem/Rating/index.ts
